findmycar/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
ebaysdk==2.2.0
SQLAlchemy==2.0.30
pydantic==2.8.2
requests==2.32.3
jinja2==3.1.4
openai==0.28.1
selenium==4.15.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
redis==5.0.0
celery==5.3.0
alembic==1.12.0
playwright==1.53.0
webdriver-manager==4.0.2
fake-useragent==1.4.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
gunicorn==21.2.0
firebase-admin==6.4.0
